% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atIndel.R
\docType{package}
\name{atIndel}
\alias{atIndel}
\title{atIndel: Affinity test for identifying regularity Indels.}
\description{
atIndel performs affinity tests of motif matches with the Indel or the
reference genomes and Indel-led changes in motif matches.
}
