% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indel_info}
\alias{indel_info}
\title{A data set for indel information.}
\format{
A list object.
}
\description{
This is a list object loaded by 'data(example)'. Each element
corresponds to an indel and includes the following components:
\tabular{ll}{
inserted_sequence \tab An integer vector representing the deroxyribose
sequence around each indel location for the longer sequence. For an
insertion, this is the inserted sequence. For an deletion, this is the
reference allele sequence. Nucleotides are coded as "A"-1, "C"-2, "G"-3,
"T"-4.\cr
insertion_len \tab The length of the inserted / deleted sequence.\cr
insertion \tab A boolean variable indicating whether the indel is an
insertion (TRUE) or a deletion (FALSE).\cr
ref \tab A character vector TODO\cr
alt \tab A character vector TODO\cr
}
}
\author{
Qinyi Zhou \email{qinyi.zhou@utdallas.edu},
Chandler Zuo \email{chandler.c.zuo@gmail.com},
Sunyoung Shin \email{sunyoung.shin@utdallas.edu}
}
