% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{indel_motif_scores}
\alias{indel_motif_scores}
\title{Compute the motif scores given a motif library and a list of indels.}
\usage{
indel_motif_scores(motif_lib, indel_info, num_cores = 1)
}
\arguments{
\item{motif_lib}{A list of the position weight matrices for the motifs.}

\item{indel_info}{A list object. Each element corresponds to an indel.
See \link{indel_info} for details and an example.}

\item{num_cores}{An integer for the number of parallel processes used for
parallel computation.}
}
\value{
A list object of position weight matrices.
}
\description{
Compute the motif scores given a motif library and a list of
}
\details{
TODO.
}
\examples{
data(example)
indel_motif_scores(motif_lib, indel_info, num_cores=1)
}
\author{
Qinyi Zhou \email{qinyi.zhou@utdallas.edu},
Chandler Zuo \email{chandler.c.zuo@gmail.com},
Sunyoung Shin \email{sunyoung.shin@utdallas.edu}
}
