% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_lib}
\alias{motif_lib}
\title{A sample motif library.}
\format{
A list object.
}
\description{
A list of the position weight matrices corresponding to motifs,
loaded by 'data(example)'.
}
\author{
Qinyi Zhou \email{qinyi.zhou@utdallas.edu},
Chandler Zuo \email{chandler.c.zuo@gmail.com},
Sunyoung Shin \email{sunyoung.shin@utdallas.edu}
}
