% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_indel_binding}
\alias{plot_indel_binding}
\title{Plot the indel binding for both reference and mutated alleles.}
\usage{
plot_indel_binding(seq_pool, seq, score, motif_pool, motif)
}
\arguments{
\item{seq_pool}{TODO}

\item{seq}{TODO}

\item{score}{TODO}

\item{motif_pool}{TODO}

\item{motif}{TODO}
}
\description{
Plot the indel binding for both reference and mutated alleles.
}
\examples{
data(example)
motif_scores <- indel_motif_scores(motif_lib, indel_info)$list
plot_indel_binding(
   indel_info,
   indel_info[1],
   motif_scores,
   motif_lib,
   motif_lib[1]
)
}
\author{
Qinyi Zhou \email{qinyi.zhou@utdallas.edu},
Sunyoung Shin \email{sunyoung.shin@utdallas.edu}
}
