% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{pvalue_single_thread_helper}
\alias{pvalue_single_thread_helper}
\title{Compute the p-values for score change for a single motif and
a set of indels with the same insertion length.}
\usage{
pvalue_single_thread_helper(
  pwm,
  motif_name,
  indel_info,
  motif_scores,
  prior,
  trans_mat,
  sample_size
)
}
\arguments{
\item{pwm}{A numeric matrix for the position weight matrix of the motif.}

\item{motif_name}{A character for the motif name.}

\item{indel_info}{A list for information related to indels. See the
argument for \code{\link{indel_p_values}} for more details.}

\item{motif_scores}{A numeric matrix including two column. The first column
is the scores for the longer sequences. The 2nd column is the scores for the
shorter sequences. Each row corresponding to one indel.}

\item{prior}{A numeric vector for the prior of the Markov Chain background
model.}

\item{trans_mat}{A numeric matrix for the transition matrix of the Markov
Chain background model.}

\item{sample_size}{An integer for the Monte-Carlo sample size.}
}
\value{
A data frame.
}
\description{
Compute the p-values for score change for a single motif and
a set of indels with the same insertion length.
}
