% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{indel_motif_scores}
\alias{indel_motif_scores}
\title{Compute the motif scores given a motif library and a list of indels.}
\usage{
indel_motif_scores(motif_lib, indel_info, num_cores = 1)
}
\arguments{
\item{motif_lib}{A list of the position weight matrices for the motifs.}

\item{indel_info}{A list object. Each element corresponds to an indel and
' includes the following components:
\tabular{ll}{
inserted_sequence \tab An integer vector representing the deroxyribose
sequence around each indel location for the longer sequence. For an
insertion, this is the inserted sequence. For an deletion, this is the
reference allele sequence. Nucleotides are coded as "A"-1, "C"-2, "G"-3,
"T"-4.\cr
insertion_len \tab The length of the inserted / deleted sequence.\cr
insertion \tab A boolean variable indicating whether the indel is an
insertion (TRUE) or a deletion (FALSE).\cr
ref \tab A character vector TODO\cr
alt \tab A character vector TODO\cr
}}

\item{num_cores}{An integer for the number of parallel processes used for
parallel computation.}
}
\value{
A list object of position weight matrices.
}
\description{
Compute the motif scores given a motif library and a list of
}
\details{
TODO.
}
\examples{
data(example)
indel_motif_scores(motif_lib, indel_info, num_cores=1)
}
\author{
Qinyi Zhou \email{qinyi.zhou@utdallas.edu},
Chandler Zuo \email{chandler.c.zuo@gmail.com},
Sunyoung Shin \email{sunyoung.shin@utdallas.edu}
}
