% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_indel_binding}
\alias{plot_indel_binding}
\title{Plot the indel binding for both reference and mutated alleles.}
\usage{
plot_indel_binding(indel_info, motif_scores, motif)
}
\arguments{
\item{indel_info}{A list of length 1 containing the indel information. See
\link{indel_info} for details and an example.}

\item{motif_scores}{A list of the same structure as the '$list' field in the
output of \code{\link{indel_motif_scores}}.}

\item{motif}{A list of length 1 containing the motif PWM. The
name of this element is the motif name.}
}
\description{
Plot the indel binding for both reference and mutated alleles.
}
\examples{
data(example)
motif_scores <- indel_motif_scores(motif_lib, indel_info)$list
plot_indel_binding(
   indel_info[1],
   motif_scores,
   motif_lib[1]
)
}
\author{
Qinyi Zhou \email{qinyi.zhou@utdallas.edu},
Sunyoung Shin \email{sunyoung.shin@utdallas.edu}
}
