% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trans_mat}
\alias{trans_mat}
\title{Default transition probability matrix for nucleotide sequences in the
reference genome.}
\format{
A 4 by 4 numeric matrix.
}
\description{
This parameter is fitted using randomly selected 100K 100bp
sequences from hg19, excluding the first and last 10k base pairs of each
chromosome, according to a 1st-order Markov model.
}
\author{
Qinyi Zhou \email{qinyi.zhou@utdallas.edu},
Chandler Zuo \email{chandler.c.zuo@gmail.com},
Sunyoung Shin \email{sunyoung.shin@utdallas.edu}
}
